+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- `ADD` Published message details can be viewed
+- `ADD` Published messages are shown in the all messages list
 - `ADD` Preview deployment cleanup supports custom branches
 - `ADD` Published messages get distributed to Kafka if enabled
 - `ADD` Messages that are approved can be published